bin - update scripts to not depend on env.sh - as the build no longer generates it

+++ bin/drop-journals.sh
@@ -1,8 +1,12 @@
 #!/bin/bash
-source $(dirname "$0")/env.sh
+if [ "$#" -ne 2 ]; then
+    echo "usage: $0 name node-list|mediator-list"
+    exit 1
+fi
 
-n=$1
-exec 4<$1
+APP_NAME=$1
+
+exec 4<$2
 while read -u4 m; do
     echo $m
     ssh -o ConnectTimeout=5 $m "sudo rm /opt/${APP_NAME}/journals/*"
